Saving token in secret manager works
